medium: further cleanup of build
