--- lib-es6/index.d.ts
@@ -1,10 +1,0 @@
-/**
- * @author Michael Breitung
- * @copyright Michael Breitung Photography (www.mibreit-photo.com)
- */
-export * from './tools/createLazyLoader';
-export { ELazyMode, LazyLoaderConfig } from './tools/createLazyLoader';
-export { default as IElementInfo } from './interfaces/IElementInfo';
-export { default as IElementLoaderInfo } from './interfaces/IElementLoaderInfo';
-export { default as ILazyLoader } from './interfaces/ILazyLoader';
-export { default as Element } from './components/Element';
