--- lib-es6/index.js
@@ -1,7 +1,0 @@
-/**
- * @author Michael Breitung
- * @copyright Michael Breitung Photography (www.mibreit-photo.com)
- */
-export * from './tools/createLazyLoader';
-export { ELazyMode } from './tools/createLazyLoader';
-export { default as Element } from './components/Element';
