+++ lib/components/ElementInfo.d.ts
