+++ lib/components/ElementInfo.js
