+++ lib/components/ElementSurrogate.d.ts
