+++ lib/components/ElementSurrogate.js
