+++ lib/components/LazyLoader.d.ts
