+++ lib/components/ScrollLoader.d.ts
