+++ lib/interfaces/IElementInfo.d.ts
