+++ lib/interfaces/IElementInfo.js
