+++ lib/interfaces/IElementLoader.js
