+++ lib/interfaces/IElementLoaderInfo.d.ts
