+++ lib/interfaces/IElementLoaderInfo.js
