+++ lib/interfaces/IElementLocationInfo.d.ts
