+++ lib/interfaces/IElementLocationInfo.js
