+++ lib/interfaces/IElementStage.d.ts
