+++ lib/interfaces/IElementStage.js
