+++ lib/interfaces/ILazyLoader.d.ts
