+++ lib/interfaces/ILazyLoader.js
