+++ lib/tools/createLazyLoader.d.ts
