+++ lib/tools/createLazyLoader.js
