+++ package.json
@@ -2,16 +2,15 @@
   "name": "mibreit-lazy-loader",
   "version": "1.2.0",
   "description": "Lazy loading of images and iframes with windowing mechanism",
-  "main": "lib/index.js",
-  "module": "lib-es6/index.js",
-  "types": "lib-es6/index.d.ts",
+  "main": "lib/index-commonjs.js",
+  "module": "lib/index.js",
+  "types": "lib/index.d.ts",
   "scripts": {
-    "start": "webpack-dev-server",
     "build": "webpack --config webpack.config.js",
     "build:dev": "webpack --config webpack.config.dev.js",
-    "prebuild:npm": "rimraf lib/ && rimraf lib-es6/",
+    "prebuild:npm": "rimraf lib/",
     "build:npm": "webpack --config webpack.config.npm.js && tsc",
-    "postbuild:npm": "copyfiles -u 1 src/**/*.svg src/**/*.css lib-es6/"
+    "postbuild:npm": "copyfiles -u 1 src/**/*.svg src/**/*.css lib/"
   },
   "keywords": [
     "Typescript"
@@ -24,10 +23,12 @@
   "devDependencies": {
     "copyfiles": "^2.4.1",
     "css-loader": "^6.7.2",
+    "css-minimizer-webpack-plugin": "^5.0.1",
+    "mini-css-extract-plugin": "^2.7.6",
+    "rimraf": "^3.0.2",
     "style-loader": "^3.3.1",
     "ts-loader": "^9.4.1",
     "typescript": "^4.9.3",
-    "rimraf": "^3.0.2",
     "typescript-plugin-css-modules": "^5.0.1",
     "webpack": "^5.88.2",
     "webpack-cli": "^4.10.0"
