+++ tsconfig.json
@@ -1,11 +1,10 @@
 {
   "compilerOptions": {
-    "outDir": "./lib-es6/",
+    "outDir": "./lib/",
     "noImplicitAny": true,
     "module": "es6",
     "target": "es5",
     "allowJs": true,
-    "declaration": true,
     "strict": true,
     "moduleResolution": "node",
     "lib": ["es6", "dom"],
